typescript setup
