+++ tsconfig.json
@@ -1,0 +1,29 @@
+{
+    "compilerOptions": {
+        "module":"es2015",
+        "target": "es2015",
+        "jsx": "react",
+        "rootDir": "src",
+        "outDir": "build",
+        "allowSyntheticDefaultImports": true,
+        "noImplicitAny": true,
+        "sourceMap": true,
+        "experimentalDecorators": true,
+        "preserveConstEnums": true,
+        "allowJs": true
+    },
+    "filesGlob": [
+        "typings/index.d.ts",
+        "src/**/*.ts",
+        "src/**/*.tsx",
+        "node_modules/typescript/lib/lib.es6.d.ts"
+    ],
+    "exclude":[
+        "index.android.js", 
+        "index.ios.js", 
+        "build", 
+        "node_modules",
+        "__tests__"
+    ],
+    "compileOnSave": false
+}
