+++ typings.json
@@ -1,0 +1,4 @@
+{
+  "name": "momentz4ever",
+  "dependencies": {}
+}
